refactor(guest-mock-codex): split app-server handlers
